split new count primitives into separate files
