+++ featuretools/primitives/standard/count_aggregation_primitives.py
@@ -5,142 +5,6 @@ from woodwork.logical_types import Integer, IntegerNullable
 from featuretools.primitives.base.aggregation_primitive_base import AggregationPrimitive
 
 
-class CountAboveMean(AggregationPrimitive):
-    """Calculates the number of values that are above the mean.
-
-    Args:
-        skipna (bool): Determines if to use NA/null values. Defaults to
-            True to skip NA/null.
-
-    Examples:
-        >>> count_above_mean = CountAboveMean()
-        >>> count_above_mean([1, 2, 3, 4, 5])
-        2
-
-        The way NaNs are treated can be controlled.
-
-        >>> count_above_mean_skipna = CountAboveMean(skipna=False)
-        >>> count_above_mean_skipna([1, 2, 3, 4, 5, None])
-        nan
-    """
-
-    name = "count_above_mean"
-    input_types = [ColumnSchema(semantic_tags={"numeric"})]
-    return_type = ColumnSchema(logical_type=IntegerNullable, semantic_tags={"numeric"})
-    stack_on_self = False
-
-    def __init__(self, skipna=True):
-        self.skipna = skipna
-
-    def get_function(self):
-        def count_above_mean(x):
-            mean = x.mean(skipna=self.skipna)
-            if np.isnan(mean):
-                return np.nan
-            return len(x[x > mean])
-
-        return count_above_mean
-
-
-class CountBelowMean(AggregationPrimitive):
-    """Determines the number of values that are below the mean.
-
-    Args:
-        skipna (bool): Determines if to use NA/null values. Defaults to
-            True to skip NA/null.
-
-    Examples:
-        >>> count_below_mean = CountBelowMean()
-        >>> count_below_mean([1, 2, 3, 4, 10])
-        3
-
-        The way NaNs are treated can be controlled.
-
-        >>> count_below_mean_skipna = CountBelowMean(skipna=False)
-        >>> count_below_mean_skipna([1, 2, 3, 4, 5, None])
-        nan
-    """
-
-    name = "count_below_mean"
-    input_types = [ColumnSchema(semantic_tags={"numeric"})]
-    return_type = ColumnSchema(logical_type=IntegerNullable, semantic_tags={"numeric"})
-    stack_on_self = False
-
-    def __init__(self, skipna=True):
-        self.skipna = skipna
-
-    def get_function(self):
-        def count_below_mean(x):
-            mean = x.mean(skipna=self.skipna)
-            if np.isnan(mean):
-                return np.nan
-            return len(x[x < mean])
-
-        return count_below_mean
-
-
-class CountGreaterThan(AggregationPrimitive):
-    """Determines the number of values greater than a controllable threshold.
-
-    Args:
-        threshold (float): The threshold to use when counting the number
-            of values greater than. Defaults to 10.
-
-    Examples:
-        >>> count_greater_than = CountGreaterThan(threshold=3)
-        >>> count_greater_than([1, 2, 3, 4, 5])
-        2
-    """
-
-    name = "count_greater_than"
-    input_types = [ColumnSchema(semantic_tags={"numeric"})]
-    return_type = ColumnSchema(logical_type=Integer, semantic_tags={"numeric"})
-    stack_on_self = False
-    default_value = 0
-
-    def __init__(self, threshold=10):
-        self.threshold = threshold
-
-    def get_function(self):
-        def count_greater_than(x):
-            return x[x > self.threshold].count()
-
-        return count_greater_than
-
-
-class CountInsideNthSTD(AggregationPrimitive):
-    """Determines the count of observations that lie inside
-        the first N standard deviations (inclusive).
-
-    Args:
-        n (float): Number of standard deviations. Default is 1
-
-    Examples:
-        >>> count_inside_nth_std = CountInsideNthSTD(n=1.5)
-        >>> count_inside_nth_std([1, 10, 15, 20, 100])
-        4
-    """
-
-    name = "count_inside_nth_std"
-    input_types = [ColumnSchema(semantic_tags={"numeric"})]
-    return_type = ColumnSchema(logical_type=Integer, semantic_tags={"numeric"})
-    stack_on_self = False
-    default_value = 0
-
-    def __init__(self, n=1):
-        if n < 0:
-            raise ValueError("n must be a positive number")
-
-        self.n = n
-
-    def get_function(self):
-        def count_inside_nth_std(x):
-            cond = np.abs(x - np.mean(x)) <= np.std(x) * self.n
-            return cond.sum()
-
-        return count_inside_nth_std
-
-
 class CountInsideRange(AggregationPrimitive):
     """Determines the number of values that fall within a certain range.
